Blanket impl ColdStore for any KeyValueStore
The two explicit impls (BeaconNodeBackend, MemoryStore) were identical
boilerplate translating slot/root keys into the underlying byte-keyed
KeyValueStore. Replace with a single blanket impl in lib.rs.

Forecloses a future ColdStore impl that isn't a KeyValueStore (e.g. wiring
StaticColdStore directly as the Cold parameter); reversible if/when that
becomes wanted.
