+++ beacon_node/store/src/memory_store.rs
@@ -1,13 +1,10 @@
 use crate::{
-    ColdStore, ColumnIter, ColumnKeyIter, DBColumn, DBColumnCold, DBColumnColdIndex, Error,
-    ItemStore, Key, KeyValueStore, KeyValueStoreOp, SlotIter, get_key_for_col,
-    hot_cold_store::BytesKey,
+    ColumnIter, ColumnKeyIter, DBColumn, Error, ItemStore, Key, KeyValueStore, KeyValueStoreOp,
+    get_key_for_col, hot_cold_store::BytesKey,
 };
 use parking_lot::RwLock;
-use ssz::{Decode, Encode};
 use std::collections::{BTreeMap, HashSet};
 use std::marker::PhantomData;
-use types::Hash256;
 use types::*;
 
 type DBMap = BTreeMap<BytesKey, Vec<u8>>;
@@ -154,67 +151,3 @@ impl<E: EthSpec> KeyValueStore<E> for MemoryStore<E> {
 }
 
 impl<E: EthSpec> ItemStore<E> for MemoryStore<E> {}
-
-impl<E: EthSpec> ColdStore<E> for MemoryStore<E> {
-    fn get(&self, column: DBColumnCold, slot: Slot) -> Result<Option<Vec<u8>>, Error> {
-        KeyValueStore::get_bytes(self, column.db_column(), &slot.as_u64().to_be_bytes())
-    }
-
-    fn put_batch(&self, column: DBColumnCold, items: Vec<(Slot, Vec<u8>)>) -> Result<(), Error> {
-        let col = column.db_column();
-        let ops = items
-            .into_iter()
-            .map(|(slot, value)| {
-                KeyValueStoreOp::PutKeyValue(col, slot.as_u64().to_be_bytes().to_vec(), value)
-            })
-            .collect();
-        KeyValueStore::do_atomically(self, ops)
-    }
-
-    fn contains(&self, column: DBColumnCold, slot: Slot) -> Result<bool, Error> {
-        KeyValueStore::key_exists(self, column.db_column(), &slot.as_u64().to_be_bytes())
-    }
-
-    fn iter_from(&self, column: DBColumnCold, from: Slot) -> SlotIter<'_> {
-        Box::new(
-            KeyValueStore::iter_column_from::<Vec<u8>>(
-                self,
-                column.db_column(),
-                &from.as_u64().to_be_bytes(),
-            )
-            .map(|res| {
-                res.and_then(|(key_bytes, value)| {
-                    let bytes: [u8; 8] = key_bytes.try_into().map_err(|_| Error::InvalidBytes)?;
-                    Ok((Slot::new(u64::from_be_bytes(bytes)), value))
-                })
-            }),
-        )
-    }
-
-    fn get_index(&self, column: DBColumnColdIndex, root: Hash256) -> Result<Option<Slot>, Error> {
-        Ok(
-            KeyValueStore::get_bytes(self, column.db_column(), root.as_slice())?
-                .map(|bytes| Slot::from_ssz_bytes(&bytes))
-                .transpose()?,
-        )
-    }
-
-    fn put_index_batch(
-        &self,
-        column: DBColumnColdIndex,
-        items: Vec<(Hash256, Slot)>,
-    ) -> Result<(), Error> {
-        let col = column.db_column();
-        let ops = items
-            .into_iter()
-            .map(|(root, slot)| {
-                KeyValueStoreOp::PutKeyValue(col, root.as_slice().to_vec(), slot.as_ssz_bytes())
-            })
-            .collect();
-        KeyValueStore::do_atomically(self, ops)
-    }
-
-    fn sync(&self) -> Result<(), Error> {
-        KeyValueStore::sync(self)
-    }
-}
